+++ app/v1/services/costumer/costumer_manager.py
@@ -4,25 +4,41 @@ from app.v1.models import user_collection
 from app.v1.utils.email import send_email, generate_otp
 from bson import ObjectId  # Import ObjectId to work with MongoDB IDs
 import bcrypt
+
 # from app.v1.utils.token import generate_jwt_token
-from fastapi import HTTPException, status, Body,Path
+from fastapi import HTTPException, status, Body, Path, Request
 from typing import Optional
 from datetime import datetime, timedelta
 from bcrypt import hashpw, gensalt
 from app.v1.schemas.costumer.costumer import UpdateCostumerRequest
 from app.v1.utils.token import get_oauth_tokens, create_access_token, create_refresh_token
+from app.v1.middleware.auth import get_current_user
+
 
 class CostumerManager:
 
-    async def create_costumer(self, create_costumer_request: User):
+    async def create_costumer(self, request: Request, token: str, create_costumer_request: User):
         try:
+            current_user = await get_current_user(request=request, token=token)
+            if not current_user:
+                raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized")
+
+            if "admin" not in [role.value for role in current_user.roles] and current_user.user_role != 2:
+                raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Forbidden")
             # User registration logic
             existing_user = await user_collection.find_one(
-                {"$or": [{"email": {"$eq": create_costumer_request.email, "$nin": [None, ""]}}, {"phone": {"$eq": create_costumer_request.phone, "$nin": [None, ""]}}]}
+                {
+                    "$or": [
+                        {"email": {"$eq": create_costumer_request.email, "$nin": [None, ""]}},
+                        {"phone": {"$eq": create_costumer_request.phone, "$nin": [None, ""]}},
+                    ]
+                }
             )
 
             if existing_user:
-                raise HTTPException(status_code=404, detail="Costumer with this email or phone already exists in the database.")
+                raise HTTPException(
+                    status_code=404, detail="Costumer with this email or phone already exists in the database."
+                )
 
             otp = generate_otp()
 
@@ -31,12 +47,13 @@ class CostumerManager:
             #     await send_email(to_email, otp)
             # if create_costumer_request.phone:
             #     to_phone = create_costumer_request.phone
-                # await send_sms(to_phone, otp)  # Uncomment this line when implementing SMS functionality
+            # await send_sms(to_phone, otp)  # Uncomment this line when implementing SMS functionality
             otp_expiration_time = datetime.utcnow() + timedelta(minutes=10)
             # create_costumer_request.otp = otp
             # create_costumer_request.otp_expiration_time = otp_expiration_time
-            # create_costumer_request.password = hashpw(create_costumer_request.password.encode('utf-8'), gensalt()).decode('utf-8') 
+            # create_costumer_request.password = hashpw(create_costumer_request.password.encode('utf-8'), gensalt()).decode('utf-8')
             create_costumer_request_dict = create_costumer_request.dict()
+            create_costumer_request_dict["created_at"] = datetime.utcnow()
             result = await user_collection.insert_one(create_costumer_request_dict)
             create_costumer_request_dict["id"] = str(result.inserted_id)  # Add `id`
             del create_costumer_request_dict["_id"]
@@ -44,14 +61,21 @@ class CostumerManager:
         except Exception as e:
             raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=str(e))
 
-
-    async def costumer_list(self, page: int, limit: int, search: str = None, role: str = "user"):
+    async def costumer_list(
+        self, request: Request, token: str, page: int, limit: int, search: str = None, role: str = "user"
+    ):
         try:
+            current_user = await get_current_user(request=request, token=token)
+            if not current_user:
+                raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized")
+
+            if "admin" not in [role.value for role in current_user.roles] and current_user.user_role != 2:
+                raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Forbidden")
             valid_roles = ["admin", "user", "vendor"]
             if role not in valid_roles:
                 raise HTTPException(
                     status_code=status.HTTP_400_BAD_REQUEST,
-                    detail=f"Invalid role: '{role}'. Valid roles are: {valid_roles}."
+                    detail=f"Invalid role: '{role}'. Valid roles are: {valid_roles}.",
                 )
 
             skip = max((page - 1) * limit, 0)
@@ -60,16 +84,13 @@ class CostumerManager:
             if search:
                 search = search.strip()
                 if not search:
-                    raise HTTPException(
-                        status_code=status.HTTP_400_BAD_REQUEST,
-                        detail="Search term cannot be empty"
-                    )
+                    raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Search term cannot be empty")
                 search_regex = {"$regex": search, "$options": "i"}
                 query["$or"] = [
                     {"first_name": search_regex},
                     {"last_name": search_regex},
                     {"email": search_regex},
-                    {"phone": search_regex}
+                    {"phone": search_regex},
                 ]
 
             # Fields to include in the result
@@ -84,19 +105,25 @@ class CostumerManager:
                 "status": 1,
                 "created_at": 1,
                 "costumer_address": 1,
-                "costumer_details": 1
+                "costumer_details": 1,
+                "created_at": 1,
             }
 
             # Fetch paginated results
             result = await user_collection.find(query, projection).skip(skip).limit(limit).to_list(length=limit)
-
             # Format costumer data
             costumer_data = []
             for costumer in result:
-                costumer["id"] = str(costumer.pop("_id"))  # Convert ObjectId to string
+                costumer["id"] = str(costumer.pop("_id"))
                 costumer["first_name"] = costumer["first_name"].capitalize()
                 costumer["last_name"] = costumer["last_name"].capitalize()
-                costumer["email"] = costumer["email"].lower()
+                costumer["email"] = costumer["email"]
+                costumer["phone"] = costumer["phone"]
+                created_at = costumer.get("created_at")
+                if isinstance(created_at, datetime):
+                    costumer["created_at"] = created_at.isoformat()
+                else:
+                    costumer["created_at"] = str(created_at)
                 costumer_data.append(costumer)
 
             # Fetch total count for the query
@@ -105,39 +132,76 @@ class CostumerManager:
             # Calculate total pages
             total_pages = (total_costumers + limit - 1) // limit
 
-            return {
-                "data": costumer_data,
-                "total_items": total_costumers,
-                "total_pages": total_pages
-            }
+            return {"data": costumer_data, "total_items": total_costumers, "total_pages": total_pages}
         except Exception as e:
             raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=str(e))
 
-    async def get_costumer(self, id: str):
+    async def get_costumer(self, request: Request, token: str, id: str):
         try:
-            # Query for costumer role and the specific user id
+            current_user = await get_current_user(request=request, token=token)
+            if not current_user:
+                raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized")
+
+            if "admin" not in [role.value for role in current_user.roles] and current_user.user_role != 2:
+                raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Forbidden")
+
+            # Query for customer role and the specific user id
             query = {"_id": ObjectId(id), "roles": {"$in": ["user"]}}
 
-            # Find the costumer by id and role
-            result = await user_collection.find_one(query)
+            # Fields to include in the result
+            projection = {
+                "first_name": 1,
+                "last_name": 1,
+                "email": 1,
+                "phone": 1,
+                "gender": 1,
+                "status": 1,
+                "costumer_address": 1,  # Include customer address
+                "costumer_details": 1,
+                "created_at": 1,
+            }
+
+            # Find the customer by id and role
+            result = await user_collection.find_one(query, projection)
+
             if not result:
-                raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Costumer not found")
-            
+                raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Customer not found")
+
             # Convert _id to string and format other fields
             result["id"] = str(result.pop("_id"))
             result["first_name"] = result["first_name"].capitalize()
             result["last_name"] = result["last_name"].capitalize()
-            result["email"] = result["email"].lower()
-            result["phone"] = result["phone"].lower()
+            result["email"] = result["email"]
+            result["phone"] = result["phone"]
             result["gender"] = result["gender"]
+            result["status"] = result["status"]
+
+            # Format created_at field
+            created_at = result.get("created_at")
+            if isinstance(created_at, datetime):
+                result["created_at"] = created_at.isoformat()  # Convert datetime to ISO 8601 string
+            else:
+                result["created_at"] = str(created_at)
+
+            # Include customer address if it exists
+            result["costumer_address"] = result.get("costumer_address", None)
+            result["costumer_details"] = result.get("costumer_details", None)
             result.pop("password", None)
-            
+
             return result
         except Exception as ex:
             raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=str(ex))
-        
-    async def update_costumer(self, id: str, update_costumer_request: UpdateCostumerRequest):
+
+    async def update_costumer(
+        self, request: Request, token: str, id: str, update_costumer_request: UpdateCostumerRequest
+    ):
         try:
+            current_user = await get_current_user(request=request, token=token)
+            if not current_user:
+                raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized")
+
+            if "admin" not in [role.value for role in current_user.roles] and current_user.user_role != 2:
+                raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Forbidden")
             # Validate costumer ID
             if not ObjectId.is_valid(id):
                 raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=f"Invalid costumer ID: '{id}'")
@@ -167,43 +231,50 @@ class CostumerManager:
                 update_data["costumer_details"] = update_costumer_request.costumer_details
             if not update_data:
                 raise HTTPException(
-                    status_code=status.HTTP_400_BAD_REQUEST,
-                    detail="No valid fields provided for update"
-                )    
-            await user_collection.update_one(
-                {"_id": ObjectId(id)},
-                {"$set": update_data}
-            )
+                    status_code=status.HTTP_400_BAD_REQUEST, detail="No valid fields provided for update"
+                )
+            await user_collection.update_one({"_id": ObjectId(id)}, {"$set": update_data})
 
             # Fetch updated costumer data
             updated_costumer = await user_collection.find_one({"_id": ObjectId(id)})
             if not updated_costumer:
                 raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Costumer not found")
 
-            # Convert _id to string and format other fields
-            updated_costumer["id"] = str(updated_costumer.pop("_id"))
-            updated_costumer["first_name"] = updated_costumer["first_name"].capitalize()
-            updated_costumer["last_name"] = updated_costumer["last_name"].capitalize()
-            updated_costumer["email"] = updated_costumer["email"].lower()
-            updated_costumer["phone"] = updated_costumer["phone"].lower()
-            updated_costumer["gender"] = updated_costumer["gender"].capitalize()
-            updated_costumer["status"] = updated_costumer["status"]
-            updated_costumer["costumer_address"] = updated_costumer["costumer_address"]
-            updated_costumer["costumer_details"] = updated_costumer["costumer_details"]
-            updated_costumer.pop("password", None)
+            response_data = {
+                "id": str(updated_costumer.get("_id")),
+                "first_name": updated_costumer.get("first_name", "").capitalize(),
+                "last_name": updated_costumer.get("last_name", "").capitalize(),
+                "email": updated_costumer.get("email", "").lower(),
+                "phone": updated_costumer.get("phone", ""),
+                "gender": updated_costumer.get("gender", "").capitalize(),
+                "status": updated_costumer.get("status", ""),
+                "costumer_address": updated_costumer.get(
+                    "costumer_address", ""
+                ),  # Default to empty string if not present
+                "costumer_details": updated_costumer.get(
+                    "costumer_details", ""
+                ),  # Default to empty string if not present
+            }
 
-            return updated_costumer
+            # Remove sensitive fields like 'password' if they exist
+            response_data.pop("password", None)
+
+            return response_data
         except Exception as e:
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail=str(e)
-            )    
+            print(e)
+            raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=str(e))
 
-    async def delete_costumer(self, id: str):
+    async def delete_costumer(self, request: Request, token: str, id: str):
         try:
+            current_user = await get_current_user(request=request, token=token)
+            if not current_user:
+                raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized")
+
+            if "admin" not in [role.value for role in current_user.roles] and current_user.user_role != 2:
+                raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Forbidden")
             result = await user_collection.delete_one({"_id": ObjectId(id)})
             if not result:
                 raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Costumer not found")
             return {"data": None}
         except Exception as ex:
-            raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=str(ex))    
+            raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=str(ex))
